+++ packages/jpgwire/src/index.ts
@@ -3,4 +3,5 @@ export * from './certs.js';
 export * from './util.js';
 export * from './metrics.js';
 export * from './pgwire_types.js';
+export * from './socket_adapter.js';
 export * from './structure_parser.js';