+++ packages/service-core-tests/src/test-utils/general-utils.ts
@@ -1,5 +1,6 @@
 import { storage, utils } from '@powersync/service-core';
 import { GetQuerierOptions, RequestParameters, SqlSyncRules } from '@powersync/service-sync-rules';
+import { versionedHydrationState } from '@powersync/service-sync-rules/src/HydrationState.js';
 import * as bson from 'bson';
 
 export const ZERO_LSN = '0/0';
@@ -25,7 +26,10 @@ export function testRules(content: string): storage.PersistedSyncRulesContent {
       return {
         id: 1,
         sync_rules: SqlSyncRules.fromYaml(content, options),
-        slot_name: 'test'
+        slot_name: 'test',
+        hydratedSyncRules() {
+          return this.sync_rules.hydrate({ hydrationState: versionedHydrationState(1) });
+        }
       };
     },
     lock() {
@@ -112,7 +116,6 @@ export function querierOptions(globalParameters: RequestParameters): GetQuerierO
   return {
     globalParameters,
     hasDefaultStreams: true,
-    streams: {},
-    bucketIdTransformer: SqlSyncRules.versionedBucketIdTransformer('1')
+    streams: {}
   };
 }