+++ packages/service-core/src/storage/SyncRulesBucketStorage.ts
@@ -1,5 +1,5 @@
 import { Logger, ObserverClient } from '@powersync/lib-services-framework';
-import { ParameterLookup, SqlSyncRules, SqliteJsonRow } from '@powersync/service-sync-rules';
+import { HydratedSyncRules, ScopedParameterLookup, SqliteJsonRow } from '@powersync/service-sync-rules';
 import * as util from '../util/util-index.js';
 import { BucketStorageBatch, FlushedResult, SaveUpdate } from './BucketStorageBatch.js';
 import { BucketStorageFactory } from './BucketStorageFactory.js';
@@ -32,7 +32,7 @@ export interface SyncRulesBucketStorage
     callback: (batch: BucketStorageBatch) => Promise<void>
   ): Promise<FlushedResult | null>;
 
-  getParsedSyncRules(options: ParseSyncRulesOptions): SqlSyncRules;
+  getParsedSyncRules(options: ParseSyncRulesOptions): HydratedSyncRules;
 
   /**
    * Terminate the sync rules.
@@ -139,7 +139,7 @@ export interface ResolveTableOptions {
   connection_tag: string;
   entity_descriptor: SourceEntityDescriptor;
 
-  sync_rules: SqlSyncRules;
+  sync_rules: HydratedSyncRules;
 }
 
 export interface ResolveTableResult {
@@ -284,7 +284,7 @@ export interface ReplicationCheckpoint {
    *
    * This gets parameter sets specific to this checkpoint.
    */
-  getParameterSets(lookups: ParameterLookup[]): Promise<SqliteJsonRow[]>;
+  getParameterSets(lookups: ScopedParameterLookup[]): Promise<SqliteJsonRow[]>;
 }
 
 export interface WatchWriteCheckpointOptions {