+++ libs/lib-mongodb/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @powersync/lib-service-mongodb
 
+## 0.6.16
+
+### Patch Changes
+
+- @powersync/lib-services-framework@0.7.14
+
+## 0.6.15
+
+### Patch Changes
+
+- @powersync/lib-services-framework@0.7.13
+
+## 0.6.14
+
+### Patch Changes
+
+- @powersync/lib-services-framework@0.7.12
+
 ## 0.6.13
 
 ### Patch Changes