+++ packages/rsocket-router/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @powersync/service-rsocket-router
 
+## 0.2.11
+
+### Patch Changes
+
+- @powersync/lib-services-framework@0.7.14
+
+## 0.2.10
+
+### Patch Changes
+
+- @powersync/lib-services-framework@0.7.13
+
+## 0.2.9
+
+### Patch Changes
+
+- @powersync/lib-services-framework@0.7.12
+
 ## 0.2.8
 
 ### Patch Changes