+++ packages/service-core-tests/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @powersync/service-core-tests
 
+## 0.12.16
+
+### Patch Changes
+
+- Updated dependencies [8fdbf8d]
+- Updated dependencies [bdfd287]
+  - @powersync/service-core@1.18.2
+  - @powersync/service-sync-rules@0.29.10
+
+## 0.12.15
+
+### Patch Changes
+
+- Updated dependencies [21b3a41]
+  - @powersync/service-sync-rules@0.29.9
+  - @powersync/service-core@1.18.1
+
+## 0.12.14
+
+### Patch Changes
+
+- Updated dependencies [dc696b1]
+- Updated dependencies [b77bb2c]
+  - @powersync/service-core@1.18.0
+
 ## 0.12.13
 
 ### Patch Changes