+++ packages/service-core/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @powersync/service-core
 
+## 1.18.2
+
+### Patch Changes
+
+- 8fdbf8d: - Rework dynamic module loading, fixing startup issues for migration and compact jobs in 1.18.0 / 1.18.1
+- Updated dependencies [bdfd287]
+  - @powersync/service-sync-rules@0.29.10
+  - @powersync/lib-services-framework@0.7.14
+  - @powersync/service-rsocket-router@0.2.11
+
+## 1.18.1
+
+### Patch Changes
+
+- Updated dependencies [21b3a41]
+  - @powersync/service-sync-rules@0.29.9
+  - @powersync/lib-services-framework@0.7.13
+  - @powersync/service-rsocket-router@0.2.10
+
+## 1.18.0
+
+### Minor Changes
+
+- b77bb2c: - First iteration of MSSQL replication using Change Data Capture (CDC).
+  - Supports resumable snapshot replication
+  - Uses CDC polling for replication
+
+### Patch Changes
+
+- dc696b1: Clear replication errors when any replication progress has been made.
+- Updated dependencies [dc696b1]
+  - @powersync/service-types@0.13.3
+  - @powersync/lib-services-framework@0.7.12
+  - @powersync/service-rsocket-router@0.2.9
+
 ## 1.17.0
 
 ### Minor Changes